pyproject.toml
[project]
name = "tcm-ai-agent"
version = "0.1.0"
description = "AIAgent for TCM syndrome differentiation from user text/images, with WeChat integration."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.27.0",
  "httpx>=0.27.0",
  "python-multipart>=0.0.9",
  "python-dotenv>=1.0.1",
  "langchain-core",
  "langchain-openai",
]

[tool.uvicorn]
host = "0.0.0.0"
port = 8000

[tool.ruff]
line-length = 100
target-version = "py311"

